+++ v1x1-common/src/main/java/tv/v1x1/common/services/spotify/dto/Error.java
@@ -6,23 +6,23 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class Error {
     @JsonProperty
-    private int status;
+    private long status;
     @JsonProperty
     private String message;
 
     public Error() {
     }
 
-    public Error(final int status, final String message) {
+    public Error(final long status, final String message) {
         this.status = status;
         this.message = message;
     }
 
-    public int getStatus() {
+    public long getStatus() {
         return status;
     }
 
-    public void setStatus(final int status) {
+    public void setStatus(final long status) {
         this.status = status;
     }
