+++ v1x1-common/src/main/java/tv/v1x1/common/services/spotify/dto/Followers.java
@@ -8,12 +8,12 @@ public class Followers {
     @JsonProperty
     private String href;
     @JsonProperty
-    private int total;
+    private long total;
 
     public Followers() {
     }
 
-    public Followers(final String href, final int total) {
+    public Followers(final String href, final long total) {
         this.href = href;
         this.total = total;
     }
@@ -26,11 +26,11 @@ public class Followers {
         this.href = href;
     }
 
-    public int getTotal() {
+    public long getTotal() {
         return total;
     }
 
-    public void setTotal(final int total) {
+    public void setTotal(final long total) {
         this.total = total;
     }
 }
