+++ v1x1-common/src/main/java/tv/v1x1/common/services/spotify/dto/Image.java
@@ -6,26 +6,26 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class Image {
     @JsonProperty
-    private int height;
+    private long height;
     @JsonProperty
     private String url;
     @JsonProperty
-    private int width;
+    private long width;
 
     public Image() {
     }
 
-    public Image(final int height, final String url, final int width) {
+    public Image(final long height, final String url, final long width) {
         this.height = height;
         this.url = url;
         this.width = width;
     }
 
-    public int getHeight() {
+    public long getHeight() {
         return height;
     }
 
-    public void setHeight(final int height) {
+    public void setHeight(final long height) {
         this.height = height;
     }
 
@@ -37,11 +37,11 @@ public class Image {
         this.url = url;
     }
 
-    public int getWidth() {
+    public long getWidth() {
         return width;
     }
 
-    public void setWidth(final int width) {
+    public void setWidth(final long width) {
         this.width = width;
     }
 }
